+++ doc/versionHistory.rst
@@ -12,7 +12,15 @@ Version History
 6.2.0
 -------------
 
-* Add optional pipeline task to fit WCS from direct detect donut catalogs.
+* Add optional pipeline task to fit WCS from direct detect donut catalogs and generate new donut catalog from reference catalogs with fit WCS.
+
+.. _lsst.ts.wep-6.1.3:
+
+-------------
+6.1.3
+-------------
+
+* Add license information to test_generateDonutDirectDetectTask.
 
 .. _lsst.ts.wep-6.1.3: